Merged in feature/notifications (pull request #26)
Feature: Add notification system

* Add notification system

Add a notification system to show messages such as 'Save successfull'
and 'Error when searching' and similar short pieces of information. The
system uses an event bus to pass the messages from the induvidual
components to the notification area. To make showing notifications as
easy as possible a useNotifications hook is available. It returns an
object which can be deconstructed to gain access to predefined types of
messages.

* Use new notifications for user handling


Approved-by: Markus Wesslén